SecurityUtils moved into Security package
